Implement the IDFT method
The main change in this PR is the creation of the IDFT method.
This method takes the frequency and tries to recover the signal
from it. There are currently some rounding errors that occur in
both DFT and IDFT that may need to be investigated more.

I also included documentation to help indicate what the functions
do, their params, and their returns.